snowflake-to-googlesheet/requirements.txt
pandas>=1.5,<2.0
numpy>=1.24,<1.25
sqlalchemy==1.4.32 #1.3.24
snowflake-sqlalchemy==1.7.3 #1.2.4
snowflake-connector-python==3.13.1